src/application/requirements.txt
Django==3.2.5
django-environ
psycopg2
psycopg2-binary
gunicorn
daphne
djangorestframework
djangorestframework-gis
channels
channels-redis
django-webpush
pylint
pylint-django
GDAL
Pillow
django-debug-toolbar
gpxpy
simplekml
faker
factory_boy
pytest-asyncio
pytest-django
pytest-cov
reverse-geocoder